fix: properly generate zsh file w/ plugins, lazyload nvm
Also added profiling support. Before, we used to load plugins with
`source` which is technically not the `zgenom` method of doing this.
Instead, we opt to use `zgenom load` and do it within the generation
phase so that we're more zgenom-native and benefit from the caching
provided.

I also moved the `nvm` logic to be lazyloaded whenever `nvm` is ran
because otherwise new shells were taking about 1.5 seconds to load.